+++ NEWS.md
@@ -22,7 +22,7 @@ CHANGES
 * `data_modify()` now recognizes `n()`, for example to create an index for data
   groups with `1:n()` (#535).
 
-* The `replacement` argument in `data_rename()` now supports curl-styled
+* The `replacement` argument in `data_rename()` now supports glue-styled
   tokens  (#563).
 
 BUG FIXES